Split out utility methods into separate spec files
The lock acquisition tests are particularly slow, and we don't modify
those methods often, so it simplifies local development to split these
out.
